refactor(tangle-shared-ui): Move HeaderCell to shared-ui
