+++ apps/tangle-dapp/src/components/ValidatorTable.tsx
@@ -1,3 +1,4 @@
+import HeaderCell from '@tangle-network/tangle-shared-ui/components/tables/HeaderCell';
 import { sortByAddressOrIdentity } from '@tangle-network/tangle-shared-ui/components/tables/utils';
 import useNetworkStore from '@tangle-network/tangle-shared-ui/context/useNetworkStore';
 import { Validator } from '@tangle-network/tangle-shared-ui/types';
@@ -25,7 +26,7 @@ import { IS_PRODUCTION_ENV } from '../constants/env';
 import { PagePath } from '../types';
 import calculateCommission from '../utils/calculateCommission';
 import filterTableRowBy from '../utils/filterTableRowBy';
-import { HeaderCell, StringCell } from './tableCells';
+import { StringCell } from './tableCells';
 import PercentageCell from './tableCells/PercentageCell';
 import TokenAmountCell from './tableCells/TokenAmountCell';
