+++ apps/tangle-dapp/src/components/tableCells/index.ts
@@ -1,2 +1,1 @@
-export { default as HeaderCell } from './HeaderCell';
 export { default as StringCell } from './StringCell';
